refactor: Remove duplicate `useInitialSystemVolume`
Closes #348
