+++ hooks/useInitialSystemVolume.ts
@@ -1,13 +1,26 @@
+import { trackCustom } from "@equinor/mad-core";
 import { useEffect, useState } from "react";
 import SystemSetting from "react-native-system-setting";
 
 import { SYSTEM_VOLUME } from "../constants/sounds";
+import { setSystemVolume } from "../utils/volume/setSystemVolume";
 
 export const useInitialSystemVolume = () => {
   const [initialSystemVolume, setInitialSystemVolume] = useState(SYSTEM_VOLUME);
 
   useEffect(() => {
-    SystemSetting.getVolume().then((volume) => setInitialSystemVolume(volume));
+    const setInitialDeviceSystemVolume = async () => {
+      await SystemSetting.getVolume()
+        .then((volume) => setInitialSystemVolume(volume))
+        .catch((error) => {
+          trackCustom("useInitialSystemVolume: Error getting system volume", {
+            error,
+          });
+        });
+    };
+    setInitialDeviceSystemVolume();
+
+    return () => setSystemVolume(initialSystemVolume);
   }, []);
 
   return { initialSystemVolume };
